Move SignatureGenerator out of module __init__.py
It's easier structurally to have the SignatureGenerator in its own Python
module. This moves it.
